+++ src/api/caidapp/templatetags/custom_tags.py
@@ -1,4 +1,7 @@
 from django import template
+import logging
+
+logger = logging.getLogger(__name__)
 register = template.Library()
 
 @register.filter